Finalizado crud basico de produtos

+++ database/migrations/2024_06_19_020934_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('titulo');
             $table->string('categoria');
-            $table->integer('preco');
+            $table->float('preco');
             $table->timestamps();
         });
     }
